Add plates responder configuration
Allows Plates to be used easily without having to implement a custom
configuration.

Also allows easier extension of the `DefaultConfigurationSet` by
including the `$classes` parameter, which is merged with the defaults.

Simplifies the configuration tests by defining a base test case that
automatically creates the injector and applies all required configs.