Consolidate HTTP exceptions
Having one class for HTTP exceptions is more predictable and can be
handled easier.

With a single class it is no longer necessary to use method checking
to apply additional modifications to the response.

By using a HTTP status code validator it is possible to provide the
HTTP status code as the exception code, without a separate method.