Add application runner
Having a simple application runner standardizes the expectations for
bootstrapping and makes it more obvious how to execute a Spark app.

Rather than hiding the application defaults in classes, this exposes
all of the configuration and middleware directly to the user, making
it far easier to add and remove things as necessary.

Finally it provides better dependency inversion by having sane defaults
that the user can override while also preserving type hinting.