Make actions first class citizens
According to the [intent of ADR][1], the Action part of the triad is
meant to be a completely custom class that provides a boundary between
HTTP and domain logic. The action is aware of the domain and knows how
to direct output through a responder.

This is not how Equip has been treating the domain. Our attempts to
make Domain pseudo-HTTP-aware with Payload and Input has not been very
successful and is probably the least concise bit of the framework.

[1]: http://pmjones.io/adr/