Delete all responders and related configuration
The ADR triad is not meant to be implemented in an agnostic way, but
rather implemented exactly as necessary for each use case. This makes
general purpose responders pointless, since the responder attached to
each action (or possibly groups of actions) will generate an exact
response based on the requirements of the use case.