Remove payload and input classes
Payload has always been a pain point in our implementation of ADR.
Instead of using a rigid payload that must be worked around, allow the
domain to return whatever format it wants and let the action and
responder determine how to apply it to the request.

Input is too generic and does not facilitate the creation of proper
value objects that can be passed from the action into the domain.

Refs #76