Make actions first class citizens (#75)
According to the [intent of ADR][1], the Action part of the triad is
meant to be a completely custom class that provides a boundary between
HTTP and domain logic. The action is aware of the domain and knows how
to direct output through a responder.

This is not how Equip has been treating the domain. Our attempts to
make Domain pseudo-HTTP-aware with Payload and Input has not been very
successful and is probably the least concise bit of the framework.

[1]: http://pmjones.io/adr/

Also includes changes from:

* Remove payload from formatting (#77)
* Restore content negotiation (#78)
* Add documentation for v3 changes (#79)
* Remove payload and input classes (#80)
* Remove content negotiation (#84)