Make ActionInterface use ServerRequestInterface (#85)
Although ActionHandler deals exclusively with ServerRequestInterface
instead of RequestInterface, ActionInterface used just
RequestInterface. This made it impossible for actions to use the
helpful features that ServerRequestInterface offers.

In particular, this bug actually rendered the example Action code in
the docs incorrect, because the variable $request (of type
RequestInterface) was passed as a parameter to input(), which expected
ServerRequestInterface.